New feature: enable/disable verbosity
You can now enable or disable verbosity of the trainer.

If you execute the __main__.py script, add the '--verbose' flag to enable verbosity.
Omit it to disable verbosity.

If you embed the TesseractTrainer into your code, the __init__ method of this
object has the new argument 'verbose'. Its default value is set to True
in the lib.defaults.py file.
Set it to True to enable verbosity and to False to disable it.

Signed-off-by: Balthazar Rouberol <rouberol.b@gmail.com>