IMPROVEMENT: the path to the 'tiffcp' binary is now an environment variable declared in defaults.py
The 'tiffcp' path was hardcoded into the multipage_tif.py, which is awkward for
the user if (s)he wants to change it.
We now define its path into the defaults.py file, as it contains default variables
that the user can change at will.

The path is no more stored as a constant but into the environment variables dict
(os.environ).
Signed-off-by: Balthazar Rouberol <rouberol.b@gmail.com>